requirements.txt
Fabric
Flask
flask-cors
Flask-SQLAlchemy
Flask-WTF
Flask-Login
HTSQL-MYSQL
mysql-python
passlib
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
itsdangerous
requests
pandas
marshmallow
cython
gevent
